[Move]Move search.go
