+++ command/file/save.go
@@ -4,14 +4,14 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"isso0424/spotify_CLI/selfmadetypes"
-	"isso0424/spotify_CLI/selfmadetypes/responsetypes"
+	"isso0424/spotify_CLI/selfmadetypes/search"
 	"isso0424/spotify_CLI/util"
 	"os"
 )
 
 var (
 	writeFile func(string, []byte, os.FileMode) error
-	loadFile  func() (playlistList []responsetypes.SearchResultItem, err error)
+	loadFile  func() (playlistList []search.SearchResultItem, err error)
 )
 
 func init() {
@@ -19,13 +19,13 @@ func init() {
 		return ioutil.WriteFile(fileName, fileDetail, permission)
 	}
 
-	loadFile = func() (playlistList []responsetypes.SearchResultItem, err error) {
+	loadFile = func() (playlistList []search.SearchResultItem, err error) {
 		return LoadPlayList()
 	}
 }
 
 // SavePlayList is function save playlist for json file
-func SavePlayList(target responsetypes.SearchResultItem) (err error) {
+func SavePlayList(target search.SearchResultItem) (err error) {
 	playlistList, err := loadFile()
 	if err != nil {
 		return
@@ -62,7 +62,7 @@ func SavePlayList(target responsetypes.SearchResultItem) (err error) {
 
 func setSavePlayList(
 	writeFileFunc func(string, []byte, os.FileMode) error,
-	loadFileFunc func() ([]responsetypes.SearchResultItem, error),
+	loadFileFunc func() ([]search.SearchResultItem, error),
 ) func() {
 	tmpWriteFile := writeFile
 	tmpLoadFile := loadFile
