+++ util/slice.go
@@ -1,13 +1,13 @@
 package util
 
 import (
-	"isso0424/spotify_CLI/selfmadetypes/responsetypes"
+	"isso0424/spotify_CLI/selfmadetypes/search"
 	"math/rand"
 	"time"
 )
 
 // Choose is function that get 1 item from slice.
-func Choose(playlists []responsetypes.SearchResultItem) responsetypes.SearchResultItem {
+func Choose(playlists []search.SearchResultItem) search.SearchResultItem {
 	rand.Seed(time.Now().UnixNano())
 	index := rand.Intn(len(playlists))
 
@@ -15,7 +15,7 @@ func Choose(playlists []responsetypes.SearchResultItem) responsetypes.SearchResu
 }
 
 // CheckDuplicateName is function that check to exist playlist name.
-func CheckDuplicateName(name string, playlistList []responsetypes.SearchResultItem) bool {
+func CheckDuplicateName(name string, playlistList []search.SearchResultItem) bool {
 	for _, content := range playlistList {
 		if content.Name == name {
 			return false
